Use more reasonable psql options for db restores
Summary:
when ON_ERROR_STOP is not set, dumps can restore even if there's a
syntax error. That's not really a sensible thing to do.

Also add --no-psqlrc because my psqlrc has stuff in it that makes psql noisy in
batch mode (\timing).

Test Plan:
Run the swh-vault tests, which are the only ones to use this fixture
in this mode

Reviewers: #reviewers!

Differential Revision: https://forge.softwareheritage.org/D307