cli/db: do not attempt to create the DB by default
this is needed to prevent the command to depend on having pg's createdb
tool installed on the machine running the 'swh db-init' command.

In a normal deployment setup, this pg tool is not expected to be present
on every machine on which one want to be able to use this 'swh db-init'
command.