Add pytest fixture to allow mocking requests with data file
Roughly, the idea is that http queries are transformed into file queries within
http context. If found, the data file is served or not (404). This translates
into query response so that the code can actually complete as usual.

For technical reasons the pytest plugin is in a dedicated non-swh module. This
is actually required for the pytest plugin loading mecanism to work without
hijacking the loading of swh packages. See [1] for more details.

[1] https://github.com/pytest-dev/pytest/issues/2042