api/serializers: Force json module use to decode requests text response
When the simplejson module is present in the Phython environment, requests will
use it to decode json but a call to response.json(cls=...) will raise an exception
as the arguments accepted by JSONDecoder are inconsistent between the standard library
json and simplejson (more details here: https://github.com/psf/requests/issues/4842).

So ensure to use standard json module for decoding response texts.