Add specific celery task arguments to metadata sent to systemd-journald
We used to log all task arguments, which was a bit too much for the logging
pipeline to handle. Now that a lot of tasks arguments have been "canonicalized",
we can go back to logging some specific arguments directly.

Use the opportunity to drop an unused argument and improve the documentation of
these logging-related functions.