setup: Separate pytest-postgresql dependency and declare it when needed
This drops the pytest-postgresql from requirements-db.txt. This is not a
required dependency for the swh.core.db module itself.

It's a requirement for the swh.core.db.pytest_plugin. This module should become
its own module as a test requirements for other swh modules (including
swh.core.db but not limited to it).

So this commit adds a dedicated requirements-db-pytestplugin.txt file with such
dependency. This opens the db-pytestplugin key within the setup.py and then
references that test dependency for the swh.core.db modules (as it's needed for
tests of the module).

Related to T2746