Replace mock_package_sql by a mock_import_swhmodule fixture
in favor of using (and thus testing) the stock get_sql_for_package()
function (on the mock module returned by mock_import_swhmodule).

This comes with a change in expected test data structure for sql
script; the `tests/data/<mod>` is now expected to store the sql script
files in a dedicated `sql/` sub directory instead of having them
directly in the `<mod>` directory so that the get_sql_for_package() can
be used in tests unmocked.