test_db/test_db_copy_to: Fix hypothesis FailedHealthCheck error
With latest version of hypothesis, this test often raises the following
error:

hypothesis.errors.FailedHealthCheck: Data generation is extremely slow ...

Setting the number of hypothesis examples for that test to a lower value
makes the error goes away.

This should also fix that test execution in the swh-environment/tests
Jenkins job which often fails.