Remove use of app.test_client() as a ctx-manager
According to
<https://github.com/pallets/flask/issues/4734#issuecomment-1204124984>,
app.test_client() should not be used as a long-lived context manager.

This causes swh-indexer's tests to break with Flask >= 2.2.0 because
some of them start threads within the lifespan of the context manager,
which causes a crash.

As far as I can tell, simply assigning app.test_client() instead of
entering does not break anything.