Fix read-access granting to the guest user in db_utils
It was missing a commit to actually be granted.
Add a test for this specific behavior, making sure the guest user has
proper permissions.

Note that this will work if and only is the guest user does not inherit
from a bunch of default permissions when created (needs a NOINHERIT).