Migrate to copier-base template
This replaces the setup.py package file with pyproject.toml (keeping
setuptools and setuptool_scm) and use an implicit namespace declaration
for the swh namespace.

Also note that due to the poor support of PEP 420 in pytest, it needs a
pair of command line arguments to pass all tests (see tox.ini).