config: keep dicts in config ordered
YAML mappings are converted in Python as dict. The order
of the keys in dicts returned by `load_from_envvar()` were
not stable from one Python run to the next.

Since Python 3.6, dicts are ordered by default, and `yaml.safe_load()`
will always return keys in the same order. The culprit was
`merge_configs` which created a set (which are unordered) of keys
to perform the merge.