cli/db: normalize handling of given package/backend/config path argument
Make all db cli commands process given arguments as similar as possible
to each other.

Generalize fallback to using the backend class (declared in the
'swh.<package>.classes' entrypoint) to implement the 'get_datastore()'
if this later is not found in the entry point module.

Also add a message in 'swh db version' if it detects the dbmodule entry
needs to be updated in the database.