cli/db: properly support module given a full module name in `swh db init` command
When the `swh db init` command is used with a "full path" module as
argument (like `swh db init storage.proxies.masking`), the db was
initialized with this full path module name instead of the (now
expected) '<package>:<cls>' syntax.

This requires a new helper function to find the registered cls from the
module full name.