db: rework the cli db tooling for winery
These changes are necessary to make the objstorage winery backend fit in
here.

The issue was in winery, the db settings are not at the top level of the
'winery' backend configuration, but under the "database" subsection,
which was an unsupported case.

These changes make it work, but for simplification purpose, we drop the
support for 'full module path' specification of the target backend to
initialize. Namely, one cannot use the pattern:

  # this will not work any more
  swh db init -d DBCNX swh.storage.backend.masking

but stick to the entry-point pattern:

  swh db init -d DBCNX storage:masking

Or use the config file based approach.