$Add utility class to download files from S3 with resumption feature
Move common code used in swh-datasets and swh-graph related to downloading
files from a directory in a S3 bucket.

The downloader implementation was also updated to support the resumption of
failed downloads as previously if an error happened when downloading a file
an exception was raised leaving the downloaded files in an incomplete and
unusable state. Moreover, when attempting to re-download the files, all
files were entirely downloaded again regardless if they were already
successfully or partially downloaded, which can be really cumbersome when
the files to download weight several terabytes.